Add files via upload
In the Day 1 of the power BI DAX exercise, I delve deeper to solve the basic measure based exercises which was base don scenarios of different datasets. I sed SUMX, AVERAGEA, AVERAGEX, COUNTROWS, Matrix visualization and Table visualization. This was based on the easy level based but I explored the data and build measures to get more insights from the data which helped in driving results.